♻️ refactor(shared): consolidate duplicate Books types per audit §2 — shared BookSummaryFields computed surface
