+++ shared/src/commonMain/kotlin/com/calypsan/listenup/client/domain/model/BookListItem.kt
@@ -8,16 +8,19 @@ import com.calypsan.listenup.client.core.Timestamp
  *
  * No `genres`, `tags`, or `allContributors` — those are detail-only and would
  * always be empty here. Use [BookDetail] when those fields are required.
+ *
+ * Deliberately kept separate from [BookDetail] (see [BookSummaryFields] for the rationale).
+ * Computed properties and [formatDuration] are supplied once via [BookSummaryFields].
  */
 data class BookListItem(
     val id: BookId,
     val title: String,
     val sortTitle: String? = null,
     val subtitle: String? = null,
-    val authors: List<BookContributor>,
-    val narrators: List<BookContributor>,
-    val duration: Long,
-    val coverPath: String?,
+    override val authors: List<BookContributor>,
+    override val narrators: List<BookContributor>,
+    override val duration: Long,
+    override val coverPath: String?,
     val coverBlurHash: String? = null,
     val dominantColor: Int? = null,
     val darkMutedColor: Int? = null,
@@ -25,7 +28,7 @@ data class BookListItem(
     val addedAt: Timestamp,
     val updatedAt: Timestamp,
     val description: String? = null,
-    val series: List<BookSeries> = emptyList(),
+    override val series: List<BookSeries> = emptyList(),
     val publishYear: Int? = null,
     val publisher: String? = null,
     val language: String? = null,
@@ -33,28 +36,4 @@ data class BookListItem(
     val asin: String? = null,
     val abridged: Boolean = false,
     val rating: Double? = null,
-) {
-    val seriesId: String? get() = series.firstOrNull()?.seriesId
-    val seriesName: String? get() = series.firstOrNull()?.seriesName
-    val seriesSequence: String? get() = series.firstOrNull()?.sequence
-
-    val hasCover: Boolean get() = coverPath != null
-
-    val fullSeriesTitle: String?
-        get() {
-            val firstSeries = series.firstOrNull() ?: return null
-            val name = firstSeries.seriesName
-            val seq = firstSeries.sequence
-            return if (seq != null && seq.isNotBlank()) "$name #$seq" else name
-        }
-
-    val authorNames: String get() = authors.joinToString(", ") { it.name }
-    val narratorNames: String get() = narrators.joinToString(", ") { it.name }
-
-    fun formatDuration(): String {
-        val totalMinutes = duration / 60_000
-        val hours = totalMinutes / 60
-        val minutes = totalMinutes % 60
-        return if (hours > 0) "${hours}h ${minutes}m" else "${minutes}m"
-    }
-}
+) : BookSummaryFields
